chore(src/util/process): factor out process class into its own file
