+++ src/library/compiler/cpp_compiler.h
@@ -12,31 +12,23 @@ Author: Jared Roesch
 #include "kernel/expr.h"
 
 namespace lean  {
-    class process {
-        std::string m_proc_name;
-        buffer<std::string> m_args;
-    public:
-        // TODO: make this cross platform
-        process(std::string exe_name);
-        process & arg(std::string arg_str);
-        void run();
-    };
 
-    class cpp_compiler {
-      buffer<std::string> m_library_paths;
-      buffer<std::string> m_include_paths;
-      buffer<std::string> m_files;
-      buffer<std::string> m_link;
+class cpp_compiler {
+    buffer<std::string> m_library_paths;
+    buffer<std::string> m_include_paths;
+    buffer<std::string> m_files;
+    buffer<std::string> m_link;
 
-      bool m_debug;
+    bool m_debug;
 
-    public:
-        cpp_compiler & link(std::string lib);
-        cpp_compiler & library_path(std::string lib_path);
-        cpp_compiler & include_path(std::string include_path);
-        cpp_compiler & debug(bool on);
-        cpp_compiler & file(std::string file_path);
-        cpp_compiler();
-        void run();
-    };
+public:
+    cpp_compiler & link(std::string lib);
+    cpp_compiler & library_path(std::string lib_path);
+    cpp_compiler & include_path(std::string include_path);
+    cpp_compiler & debug(bool on);
+    cpp_compiler & file(std::string file_path);
+    cpp_compiler();
+    void run();
+};
+
 }
