+++ src/util/CMakeLists.txt
@@ -4,4 +4,4 @@ add_library(util OBJECT debug.cpp name.cpp name_set.cpp fresh_name.cpp
   stackinfo.cpp lean_path.cpp serializer.cpp lbool.cpp
   bitap_fuzzy_search.cpp init_module.cpp thread.cpp memory_pool.cpp
   utf8.cpp name_map.cpp list_fn.cpp null_ostream.cpp file_lock.cpp
-  small_object_allocator.cpp dynamic_library.cpp)
+  small_object_allocator.cpp dynamic_library.cpp process.cpp)
